Created initial Rust project
Initialized Cargo project, introduced some basic structures upon which
the project shall be built.

The architecture of the game follows the bottom-up approach, with the
moving force behind the development process being the set of structures
that represent the state of the program. Then a set of functions to
manipulate them is implemented.

The game has a top-level state represented by the ProgramState
structure, and then it has a set of separate modules which accept
ProgramState and return a modified one. These modules work as screen
views of the game. For example, there shall be a MenuScreen module which
allows to modify settings and clear score data. When a new game is
started, it passes the changed ProgramState to GameScreen.