Implemented Display trait for Labyrinth
It is now functionally equivalent to the implementation of Debug trait;
the difference is that it uses indexes instead of iterators to address
cells. It is beneficial since for proper display it shall be necessary
to expect the cells surrounding each one to determine what type of wall
it is: it can be an angle, a horisontal or a vertical wall et cetera.