Abstracted filtering the surrounding branches
Added function `filter_around` which accepts a field, a description of
how to treat its boundaries, the required state, and a point around
which to perform filtering, and for each surrounding cell that's
eligible calls the passed function.

Rewritten empty_neighbors using the new function.

Also `true` now represents the wall, and `false` represents the floor.