Drawing empty spaces as dots now
It used to be a space, which is a problem. With spaces it's difficult to
get a sense of scale in case of long corridors.

We've had to modify the code that displays the characters because there
are two modes of terminal display, and in the one used to render
box-drawing characters the dot is a down arrow.