Specified the Screen trait
Screen views are among the most important structures of the game. The
Screen trait is a Controller and Screen instances are Models in terms of
MVC paradigm. Once all signals to screens are described and have a
well-defined meaning, it's a matter of simple wrapping to create views
for them.