Implemented example runner of screens
The Screens modify the program state and pass it to each other, but
there must exist some global manager which switches the screens: a
screen can't replace itself. Certainly, there could exist a manager
screen that stores the child screen, dispatches the queries to it, and
performs the switching, but there's no need yet for screen nesting, so
we simply run a loop in main().