Major refactoring; replace Screen with Tickable
There are many more entities that accept user input and perform some
actions depending on it than there are screens. For example, in the menu
screen there can be a sample labyrinth with an AI solving it, so the
labyrinth itself can be tickable, while being embedded into a tickable
screen.